+++ Observe/LookDirection.cs
@@ -1,0 +1,9 @@
+﻿namespace Marioalexsan.Observe;
+
+public enum LookDirection : byte
+{
+    Default = 0,
+    Camera = 1,
+    Pose = 2,
+    Environment = 3
+}