+++ Observe/Plugin.cs
@@ -1,9 +1,8 @@
-﻿using System.Buffers.Binary;
-using System.Runtime.CompilerServices;
-using System.Text;
+﻿using System.Runtime.CompilerServices;
 using BepInEx;
 using BepInEx.Bootstrap;
 using BepInEx.Configuration;
+using BepInEx.Logging;
 using CodeTalker.Networking;
 using CodeTalker.Packets;
 using HarmonyLib;
@@ -15,22 +14,29 @@ namespace Marioalexsan.Observe;
 [BepInPlugin(ModInfo.GUID, ModInfo.NAME, ModInfo.VERSION)]
 [BepInDependency("EasySettings", BepInDependency.DependencyFlags.SoftDependency)]
 [BepInDependency("CodeTalker")]
+[BepInDependency("Soggy_Pancake.CommandLib")]
 public class ObservePlugin : BaseUnityPlugin
 {
     private struct PlayerLookInfo
     {
-        public Quaternion CameraRotation;
+        public Quaternion DesiredLookDirection;
         public PlayerRaceModel RaceModel;
         public Transform Head;
         public Quaternion LastHeadRotation;
-        public bool IgnoreCamera;
+        public bool OwlMode;
+        public LookSpeed LookSpeed;
+        public LookDirection OverrideDirection; // Only used for secondary behaviour!
+        public bool VanillaMode;
     }
 
     private struct SentPlayerInfo
     {
-        public Quaternion CameraRotation;
-        public bool IgnoreCamera;
+        public Quaternion DesiredLookDirection;
+        public bool VanillaMode;
+        public bool OwlMode;
+        public LookSpeed LookSpeed;
         public DateTime Timestamp;
+        public LookDirection OverrideDirection; // Only used for secondary behaviour!
     }
     
     private readonly Harmony _harmony = new Harmony($"{ModInfo.GUID}");
@@ -39,9 +45,14 @@ public class ObservePlugin : BaseUnityPlugin
     private readonly Dictionary<ulong, SentPlayerInfo> MultiplayerData = [];
     internal static HashSet<Player> Players { get; } = [];
     
-    private static ConfigEntry<bool> Enabled = null!;
-    private static ConfigEntry<bool> IgnoreCameraSetting = null!;
-    private static ConfigEntry<bool> EnableNetworking = null!;
+    internal static ConfigEntry<bool> Enabled = null!;
+    internal static ConfigEntry<bool> VanillaModeSetting = null!;
+    internal static ConfigEntry<bool> EnableNetworking = null!;
+    internal static ConfigEntry<bool> OwlModeSetting = null!;
+    internal static ConfigEntry<bool> AllowOwlModeForOthers = null!;
+    internal static ConfigEntry<LookSpeed> LookSpeedSetting = null!;
+    internal static ConfigEntry<bool> HoldHeadDirectionAfterStrafing = null!;
+    internal static ConfigEntry<float> HoldHeadDirectionAfterStrafingDuration = null!;
 
     private bool _lastSendEnableState = false;
     private bool _shouldSendCurrentState = false;
@@ -49,11 +60,23 @@ public class ObservePlugin : BaseUnityPlugin
     private static TimeSpan RefreshDirectionAccumulator = TimeSpan.Zero;
     private static TimeSpan PacketSendCooldown = TimeSpan.Zero;
 
+    internal static LookDirection LocalOverrideDirection = LookDirection.Default;
+    internal static TimeSpan LocalOverrideDirectionTime = TimeSpan.Zero;
+    internal static Quaternion SavedOverride = Quaternion.identity;
+
+    internal new static ManualLogSource Logger = null!;
+
     public ObservePlugin()
     {
+        Logger = base.Logger;
         Enabled = Config.Bind("General", "Enabled", true, "Enable or disable mod functionality completely.");
-        IgnoreCameraSetting = Config.Bind("General", "IgnoreCamera", false, "While active, makes your player look forward instead of using the camera direction.");
+        VanillaModeSetting = Config.Bind("General", "VanillaMode", false, "While active, makes your character's head act the same as in vanilla.");
         EnableNetworking = Config.Bind("General", "EnableNetworking", true, "Enable sending/receiving camera directions to/from people with the mod installed.");
+        OwlModeSetting = Config.Bind("Display", "OwlMode", false, "Enables full range of rotation for your player's head. This might look pretty weird in some cases!");
+        AllowOwlModeForOthers = Config.Bind("Display", "AllowOwlModeForOthers", true, "Allow other players that use OwlMode to display their unconstrained head rotations.");
+        LookSpeedSetting = Config.Bind("Display", "LookSpeed", LookSpeed.Normal, "The speed at which your character reacts to changes in direction.");
+        HoldHeadDirectionAfterStrafing = Config.Bind("Controls", "HoldHeadDirectionAfterStrafing", true, "Enable to keep looking at the given direction after strafing as if you used \"/observe environment\".");
+        HoldHeadDirectionAfterStrafingDuration = Config.Bind("Controls", "HoldHeadDirectionAfterStrafingDuration", 4f, new ConfigDescription("How long to continue looking at the environment when HoldHeadDirectionAfterStrafing is enabled, in seconds.", new AcceptableValueRange<float>(0, 120)));
     }
 
     private void Awake()
@@ -63,23 +86,70 @@ public class ObservePlugin : BaseUnityPlugin
         if (Chainloader.PluginInfos.ContainsKey("EasySettings"))
             RegisterEasySettings();
 
-        CodeTalkerNetwork.RegisterBinaryListener<LookPacket>(HandleNetworkSyncFromOthers);
+        // TODO: Implement and test as soft dependency
+        if (Chainloader.PluginInfos.ContainsKey("CodeTalker"))
+            RegisterCodeYapperListener();
+
+        // TODO: Implement and test as soft dependency
+        if (Chainloader.PluginInfos.ContainsKey("Soggy_Pancake.CommandLib"))
+            RegisterCommandLibCommands();
+    }
+
+    [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+    private void RegisterCodeYapperListener()
+    {
+        try
+        {
+            CodeTalkerNetwork.RegisterBinaryListener<LookPacket>(HandleNetworkSyncFromOthers);
+        }
+        catch (Exception e)
+        {
+            Logger.LogWarning("Failed to register network listener for CodeYapper! Please report this to the mod author!");
+            Logger.LogWarning($"Exception: {e}");
+        }
+    }
+
+    [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+    private void RegisterCommandLibCommands()
+    {
+        try
+        {
+            Commands.RegisterCommands();
+        }
+        catch (Exception e)
+        {
+            Logger.LogWarning("Failed to register commands for CommandLib! Please report this to the mod author!");
+            Logger.LogWarning($"Exception: {e}");
+        }
     }
 
     [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
     private void RegisterEasySettings()
     {
-        Settings.OnInitialized.AddListener(() =>
+        try
         {
-            Settings.ModTab.AddHeader("Observe");
-            Settings.ModTab.AddToggle("Enabled", Enabled);
-            Settings.ModTab.AddToggle("Ignore Camera (self)", IgnoreCameraSetting);
-            Settings.ModTab.AddToggle("Enable Networking", EnableNetworking);
-        });
-        Settings.OnApplySettings.AddListener(() =>
+            Settings.OnInitialized.AddListener(() =>
+            {
+                Settings.ModTab.AddHeader("Observe");
+                Settings.ModTab.AddToggle("Enabled", Enabled);
+                Settings.ModTab.AddToggle("Ignore Camera (self)", VanillaModeSetting);
+                Settings.ModTab.AddToggle("Enable Networking", EnableNetworking);
+                Settings.ModTab.AddToggle("Owl Mode (full rotations)", OwlModeSetting);
+                Settings.ModTab.AddToggle("Allow Owl Mode for others", AllowOwlModeForOthers);
+                Settings.ModTab.AddDropdown("Look speed", LookSpeedSetting);
+                Settings.ModTab.AddToggle("Hold head direction after strafing", HoldHeadDirectionAfterStrafing);
+                Settings.ModTab.AddAdvancedSlider("Strafe hold head direction duration", HoldHeadDirectionAfterStrafingDuration, true);
+            });
+            Settings.OnApplySettings.AddListener(() =>
+            {
+                Config.Save();
+            });
+        }
+        catch (Exception e)
         {
-            Config.Save();
-        });
+            Logger.LogWarning("Failed to register settings for EasySettings! Please report this to the mod author!");
+            Logger.LogWarning($"Exception: {e}");
+        }
     }
 
     private void HandleNetworkSyncFromOthers(PacketHeader header, BinaryPacketBase packet)
@@ -108,20 +178,30 @@ public class ObservePlugin : BaseUnityPlugin
         if (!validPlayer)
             return; // Junk
 
-        Logger.LogInfo("Received packet");
         MultiplayerData[lookPacket.TargetNetId] = new SentPlayerInfo()
         {
-            CameraRotation = lookPacket.CameraRotation,
-            IgnoreCamera = lookPacket.IgnoreCamera,
             Timestamp = DateTime.UtcNow,
+            DesiredLookDirection = lookPacket.CameraRotation,
+            VanillaMode = lookPacket.VanillaMode,
+            OwlMode = lookPacket.OwlMode,
+            OverrideDirection = lookPacket.OverrideDirection,
+            LookSpeed = lookPacket.LookSpeed,
         };
     }
 
     private void LateUpdate()
     {
+        // Handle override timer
+        LocalOverrideDirectionTime -= TimeSpan.FromSeconds(Time.deltaTime);
+        if (LocalOverrideDirectionTime < TimeSpan.Zero)
+        {
+            LocalOverrideDirectionTime = TimeSpan.Zero;
+            LocalOverrideDirection = LookDirection.Default;
+        }
+        
         HandlePlayerCleanup();
-        HandleNetworkSyncToOthers();
-        HandleLocalPlayers();
+        HandleLocalPlayerAndNetworkSyncToOthers();
+        HandleOtherPlayers();
     }
 
     private void HandlePlayerCleanup()
@@ -138,7 +218,7 @@ public class ObservePlugin : BaseUnityPlugin
         Players.RemoveWhere(player => !player);
     }
 
-    private void HandleNetworkSyncToOthers()
+    private void HandleLocalPlayerAndNetworkSyncToOthers()
     {
         if (Player._mainPlayer && _cachedPlayerData.TryGetValue(Player._mainPlayer, out var mainPlayerLookData))
         {
@@ -148,13 +228,13 @@ public class ObservePlugin : BaseUnityPlugin
 
             if (_lastSendEnableState || EnableNetworking.Value)
             {
-                bool ignoreCameraSwitched = mainPlayerLookData.IgnoreCamera != IgnoreCameraSetting.Value;
+                bool ignoreCameraSwitched = mainPlayerLookData.VanillaMode != VanillaModeSetting.Value;
             
-                if (IgnoreCameraSetting.Value)
+                if (VanillaModeSetting.Value)
                     RefreshDirectionAccumulator -= TimeSpan.FromSeconds(Time.deltaTime);
                 else
                 {
-                    var angleChange = Quaternion.Angle(mainPlayerLookData.CameraRotation, CameraFunction._current.transform.rotation);
+                    var angleChange = Quaternion.Angle(mainPlayerLookData.DesiredLookDirection, CameraFunction._current.transform.rotation);
                     RefreshDirectionAccumulator -= TimeSpan.FromSeconds(Time.deltaTime + angleChange * SecondsPerSend / DegreesPerSend);
                 }
             
@@ -169,9 +249,20 @@ public class ObservePlugin : BaseUnityPlugin
                 RefreshDirectionAccumulator = TimeSpan.Zero;
             }
             
-            // Add local settings to cache
-            mainPlayerLookData.CameraRotation = CameraFunction._current.transform.rotation;
-            mainPlayerLookData.IgnoreCamera = IgnoreCameraSetting.Value;
+            // Check if we need to pose the player with strafe
+            if (Player._mainPlayer._pMove._strafeToggle && HoldHeadDirectionAfterStrafing.Value)
+            {
+                LocalOverrideDirectionTime = TimeSpan.FromSeconds(HoldHeadDirectionAfterStrafingDuration.Value);
+                LocalOverrideDirection = LookDirection.Environment;
+                SavedOverride = CameraFunction._current.transform.rotation;
+            }
+            
+            // Add settings for local player to cache
+            mainPlayerLookData.DesiredLookDirection = GetMainPlayerTargetRotation(mainPlayerLookData.Head);
+            mainPlayerLookData.VanillaMode = VanillaModeSetting.Value;
+            mainPlayerLookData.OwlMode = OwlModeSetting.Value;
+            mainPlayerLookData.LookSpeed = LookSpeedSetting.Value;
+            mainPlayerLookData.OverrideDirection = LocalOverrideDirection;
             _cachedPlayerData[Player._mainPlayer] = mainPlayerLookData;
         }
 
@@ -186,22 +277,26 @@ public class ObservePlugin : BaseUnityPlugin
         {
             _shouldSendCurrentState = false;
             PacketSendCooldown = TimeSpan.FromMilliseconds(250);
-            
-            foreach (var player in Players)
+
+            if (Player._mainPlayer && _cachedPlayerData.TryGetValue(Player._mainPlayer, out var mainPlayerData))
             {
-                if (player != Player._mainPlayer)
+                LookPacket.Instance.CameraRotation = VanillaModeSetting.Value ? Quaternion.identity : GetMainPlayerTargetRotation(mainPlayerData.Head);
+                LookPacket.Instance.TargetNetId = Player._mainPlayer.netId;
+                LookPacket.Instance.VanillaMode = VanillaModeSetting.Value;
+                LookPacket.Instance.OwlMode = OwlModeSetting.Value;
+                LookPacket.Instance.LookSpeed = LookSpeedSetting.Value;
+                LookPacket.Instance.OverrideDirection = LocalOverrideDirection;
+            
+                foreach (var player in Players)
                 {
-                    LookPacket.Instance.CameraRotation = IgnoreCameraSetting.Value ? Quaternion.identity : CameraFunction._current.transform.rotation;
-                    LookPacket.Instance.TargetNetId = Player._mainPlayer.netId;
-                    LookPacket.Instance.IgnoreCamera = IgnoreCameraSetting.Value;
-                    Logger.LogInfo("Sent packet");
-                    CodeTalkerNetwork.SendNetworkPacket(player, LookPacket.Instance);
+                    if (player != Player._mainPlayer)
+                        CodeTalkerNetwork.SendNetworkPacket(player, LookPacket.Instance);
                 }
             }
         }
     }
 
-    private void HandleLocalPlayers()
+    private void HandleOtherPlayers()
     {
         foreach (var player in Players)
         {
@@ -211,12 +306,15 @@ public class ObservePlugin : BaseUnityPlugin
                 // If so, assume they disabled sending info and make them act as vanilla
                 if (sentData.Timestamp + TimeSpan.FromSeconds(6) < DateTime.UtcNow)
                 {
-                    sentData.IgnoreCamera = true;
+                    sentData.VanillaMode = true;
                     MultiplayerData[player.netId] = sentData;
                 }
                 
-                lookData.CameraRotation = sentData.CameraRotation;
-                lookData.IgnoreCamera = sentData.IgnoreCamera;
+                lookData.DesiredLookDirection = sentData.DesiredLookDirection;
+                lookData.VanillaMode = sentData.VanillaMode;
+                lookData.OwlMode = AllowOwlModeForOthers.Value && sentData.OwlMode;
+                lookData.LookSpeed = sentData.LookSpeed;
+                lookData.OverrideDirection = sentData.OverrideDirection;
                 _cachedPlayerData[player] = lookData;
             }
             
@@ -237,26 +335,20 @@ public class ObservePlugin : BaseUnityPlugin
 
             if (raceModel != null)
             {
-                var head = raceModel._armatureTransform
-                    ?.Find("Armature_character")
-                    ?.Find("masterBone")
-                    ?.Find("hipCtrl")
-                    ?.Find("hip")
-                    ?.Find("lowBody")
-                    ?.Find("midBody")
-                    ?.Find("torso")
-                    ?.Find("neck")
-                    ?.Find("head");
+                var head = GetHeadBone(raceModel);
 
                 if (head != null)
                 {
                     _cachedPlayerData[player] = data = new PlayerLookInfo()
                     {
-                        CameraRotation = player.transform.rotation,
+                        DesiredLookDirection = player.transform.rotation,
                         RaceModel = raceModel,
                         Head = head,
                         LastHeadRotation = Quaternion.identity,
-                        IgnoreCamera = false
+                        VanillaMode = false,
+                        OwlMode = false,
+                        OverrideDirection = LookDirection.Default,
+                        LookSpeed = LookSpeed.Normal,
                     };
                 }
             }
@@ -265,44 +357,104 @@ public class ObservePlugin : BaseUnityPlugin
         if (data.Head == null)
             return;
 
-        if (data.IgnoreCamera)
+        if (data.VanillaMode)
         {
             data.LastHeadRotation = Quaternion.identity;
-            data.CameraRotation = player.transform.rotation;
+            data.DesiredLookDirection = player.transform.rotation;
             _cachedPlayerData[player] = data;
             return;
         }
 
         var playerTransform = player.transform;
-        var cameraRotation = data.CameraRotation;
+        bool isMirrored = player._pVisual._playerAppearanceStruct._isLeftHanded;
 
-        var cameraAngle = Quaternion.Angle(playerTransform.rotation, cameraRotation);
+        var targetLookRotation = data.DesiredLookDirection;
+        var targetAngle = Quaternion.Angle(playerTransform.rotation, targetLookRotation);
 
         const float MaxHeadRotation = 70f;
 
-        bool farLook = cameraAngle >= 105;
-        bool reallyFarLook = cameraAngle >= 160;
+        bool farLook = targetAngle >= 105;
+        bool reallyFarLook = targetAngle >= 160;
 
-        // Prefer looking over the right shoulder if camera is towards the back
-        var cameraRotationToConsider = !reallyFarLook ? cameraRotation : Quaternion.LookRotation(playerTransform.right, playerTransform.up);
-        Quaternion wantsToLookAt = Quaternion.RotateTowards(data.Head.parent.rotation, cameraRotationToConsider, MaxHeadRotation);
-        var currentHeadRotation = Quaternion.Slerp(data.LastHeadRotation, wantsToLookAt, Time.deltaTime * 3.5f);
+        // When it's diametrally opposite from the forward direction and is using default mode,
+        // opt to instead act as /observe camera
+        var wantsToLookAt = data.OverrideDirection != LookDirection.Default || !reallyFarLook || data.OwlMode ? targetLookRotation : GetDefaultModeBackwardLookDirection(data, player);
+        
+        Quaternion atBestCanLookAt = data.OwlMode ? wantsToLookAt : Quaternion.RotateTowards(data.Head.parent.rotation, wantsToLookAt, MaxHeadRotation);
+        var currentHeadRotation = Quaternion.Slerp(data.LastHeadRotation, atBestCanLookAt, Time.deltaTime * data.LookSpeed.MapToMultiplier());
         
         // Add a clamp in case we have a sudden direction change that would make the head do "funny" movements
-        currentHeadRotation = Quaternion.RotateTowards(data.Head.parent.rotation, currentHeadRotation, MaxHeadRotation);
+        if (!data.OwlMode)
+            currentHeadRotation = Quaternion.RotateTowards(data.Head.parent.rotation, currentHeadRotation, MaxHeadRotation);
         
         data.Head.rotation = currentHeadRotation;
         data.LastHeadRotation = data.Head.rotation;
+        
+        // If the character is mirrored, we need to also mirror the rotation on the X axis for it to be correct
+        if (isMirrored)
+        {
+            var rot = data.Head.rotation;
+            data.Head.rotation.Set(rot.x, -rot.y, -rot.z, rot.w);
+        }
 
-        if (farLook)
+        if (!data.OwlMode && data.OverrideDirection == LookDirection.Default && farLook && !reallyFarLook)
         {
             if (data.RaceModel._currentEyeCondition == EyeCondition.Center)
             {
-                float lookRightAngle = Quaternion.Angle(Quaternion.LookRotation(playerTransform.right * -1, playerTransform.up), cameraRotationToConsider);
-                data.RaceModel.Set_EyeCondition(lookRightAngle <= 90 ? EyeCondition.Right : EyeCondition.Left, 0.15f);
+                float lookRightAngle = Quaternion.Angle(Quaternion.LookRotation(playerTransform.right * -1, playerTransform.up), wantsToLookAt);
+                // If mirrored, need to reverse horizontal eye direction too
+                data.RaceModel.Set_EyeCondition((lookRightAngle <= 90) != isMirrored ? EyeCondition.Right : EyeCondition.Left, 0.15f);
             }
         }
 
         _cachedPlayerData[player] = data;
+    }
+
+    private static Quaternion GetMainPlayerTargetRotation(Transform headTransform)
+    {
+        var player = Player._mainPlayer;
+
+        bool cameraDisabled = CameraChecks.IsUsingFreecam() || (DialogManager._current && DialogManager._current._isDialogEnabled);
+        
+        switch (LocalOverrideDirection)
+        {
+            case LookDirection.Camera:
+                if (cameraDisabled)
+                    return headTransform.rotation;
+                return Quaternion.LookRotation(-CameraFunction._current.transform.forward, player.transform.up);
+            case LookDirection.Pose:
+                return headTransform.parent.rotation * SavedOverride;
+            case LookDirection.Environment:
+                return SavedOverride;
+            default:
+                if (cameraDisabled)
+                    return headTransform.rotation;
+                // Look in line with camera if no overrides are specified
+                return CameraFunction._current.transform.rotation;
+        }
+    }
+
+    private static Quaternion GetDefaultModeBackwardLookDirection(PlayerLookInfo data, Player player)
+    {
+        bool lookAtYourself = false;
+        
+        if (lookAtYourself)
+            return Quaternion.LookRotation(-(data.DesiredLookDirection * Vector3.forward), player.transform.up);
+        
+        return player.transform.rotation * Quaternion.Euler(0, 90, 0);
+    }
+
+    internal static Transform? GetHeadBone(PlayerRaceModel raceModel)
+    {
+        return raceModel._armatureTransform
+            ?.Find("Armature_character")
+            ?.Find("masterBone")
+            ?.Find("hipCtrl")
+            ?.Find("hip")
+            ?.Find("lowBody")
+            ?.Find("midBody")
+            ?.Find("torso")
+            ?.Find("neck")
+            ?.Find("head");
     }
 }